Fix pnpm before installing. Use pnpm to uninstall.
