+++ packages/core/src/extensions/install-extension/install-extension.injectable.ts
@@ -5,114 +5,28 @@
  */
 
 import { prefixedLoggerInjectable } from "@freelensapp/logger";
-import { isErrnoException } from "@freelensapp/utilities";
 import { getInjectable } from "@ogre-tools/injectable";
 import AwaitLock from "await-lock";
-import { fork } from "child_process";
-import { once } from "lodash";
-import directoryForUserDataInjectable from "../../common/app-paths/directory-for-user-data/directory-for-user-data.injectable";
-import pathToPnpmCliInjectable from "../../common/app-paths/path-to-pnpm-cli.injectable";
-import readJsonFileInjectable from "../../common/fs/read-json-file.injectable";
-import writeJsonFileInjectable from "../../common/fs/write-json-file.injectable";
-import joinPathsInjectable from "../../common/path/join-paths.injectable";
-import extensionPackageRootDirectoryInjectable from "./extension-package-root-directory.injectable";
-
-import type { PackageJson } from "../common-api";
-
-const basePnpmInstallArgs = ["install", "--prefer-offline", "--prod", "--save-optional"];
+import forkPnpmInjectable from "./fork-pnpm.injectable";
 
 export type InstallExtension = (name: string) => Promise<void>;
 
 const installExtensionInjectable = getInjectable({
   id: "install-extension",
   instantiate: (di): InstallExtension => {
-    const directoryForUserData = di.inject(directoryForUserDataInjectable);
-    const pathToPnpmCli = di.inject(pathToPnpmCliInjectable);
-    const extensionPackageRootDirectory = di.inject(extensionPackageRootDirectoryInjectable);
-    const readJsonFile = di.inject(readJsonFileInjectable);
-    const writeJsonFile = di.inject(writeJsonFileInjectable);
-    const joinPaths = di.inject(joinPathsInjectable);
     const logger = di.inject(prefixedLoggerInjectable, "EXTENSION-INSTALLER");
 
-    const pnpmHome = joinPaths(directoryForUserData, "pnpm", "data");
-    const cacheDir = joinPaths(directoryForUserData, "cache");
-    const configDir = joinPaths(directoryForUserData, "config");
-    const stateDir = joinPaths(directoryForUserData, "state");
-
-    const forkPnpm = (...args: string[]) =>
-      new Promise<void>((resolve, reject) => {
-        const child = fork(pathToPnpmCli, args, {
-          cwd: extensionPackageRootDirectory,
-          silent: false,
-          env: {
-            ...process.env,
-            PNPM_HOME: pnpmHome,
-            XDG_CACHE_HOME: cacheDir,
-            XDG_CONFIG_HOME: configDir,
-            XDG_STATE_HOME: stateDir,
-          },
-        });
-        let stdout = "";
-        let stderr = "";
-
-        child.stdout?.on("data", (data) => {
-          stdout += String(data);
-        });
-
-        child.stderr?.on("data", (data) => {
-          stderr += String(data);
-        });
-
-        child.on("close", (code) => {
-          if (code !== 0) {
-            reject(new Error([stdout, stderr].join("\n")));
-          } else {
-            resolve();
-          }
-        });
-
-        child.on("error", (error) => {
-          reject(error);
-        });
-      });
-
-    const packageJsonPath = joinPaths(extensionPackageRootDirectory, "package.json");
-
-    /**
-     * NOTES:
-     *   - We have to keep the `package.json` because `pnpm install` removes files from `node_modules`
-     *     if they are no longer in the `package.json`
-     *   - In v6.2.X we saved bundled extensions as `"dependencies"` and external extensions as
-     *     `"optionalDependencies"` at startup. This was done because `"optionalDependencies"` can
-     *     fail to install and that is OK.
-     *   - We continue to maintain this behavior here by only installing new dependencies as
-     *     `"optionalDependencies"`
-     */
-    const fixupPackageJson = once(async () => {
-      try {
-        const packageJson = (await readJsonFile(packageJsonPath)) as PackageJson;
-
-        delete packageJson.dependencies;
-
-        await writeJsonFile(packageJsonPath, packageJson);
-      } catch (error) {
-        if (isErrnoException(error) && error.code === "ENOENT") {
-          return;
-        }
-
-        throw error;
-      }
-    });
+    const forkPnpm = di.inject(forkPnpmInjectable);
 
     const installLock = new AwaitLock();
 
     return async (name) => {
       await installLock.acquireAsync();
-      await fixupPackageJson();
 
       try {
         logger.info(`installing package for extension "${name}"`);
-        await forkPnpm(...basePnpmInstallArgs, name);
+        await forkPnpm("install", "--prefer-offline", "--prod", "--force");
+        await forkPnpm("install", "--prefer-offline", "--prod", "--force", "--save-optional", name);
         logger.info(`installed package for extension "${name}"`);
       } finally {
         installLock.release();
